# src/main/resources/openAPI/parser-api.openapi.yml
openapi: 3.0.0
info:
  title: Data Harvest Parser
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: Development Server

paths:
  /user/{id}:
    get:
      summary: User Info
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      operationId: showUserInfo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"

  /auth/signup:
    post:
      summary: Register
      tags:
        - Authorization
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequestOpenApi'
      responses:
        '201':
          description: User registered successfully and waits for email verification!
        '400':
          description: Account with such email is already in use!
        '409':
          description: Account with such username is already in use!

  /auth/signin:
    post:
      summary: Login
      tags:
        - Authorization
      operationId: authenticateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestOpenApi"
      responses:
        '200':
          description: Successfully Logged In
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtResponseOpenApi"
        '400':
          description: Authentication error! Incorrect login/password input.

  /auth/activation:
    patch:
      summary: Activation User account via Email
      tags:
        - Authorization
      operationId: activateUser
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The User crypto ID
      responses:
        '200':
          description: Account was confirmed successfully

        '400':
          description: Account is already confirmed!

        '404':
          description: The link isn't valid. User with such user id not found.


  /parser/settings:
    post:
      security:
        - bearerAuth: [ ]
      summary: Set Settings for Parsing
      tags:
        - Parser
      operationId: setParserSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserParserSettingsOpenApi"
      responses:
        '201':
          description: Settings were Successfully Set
        '401':
          $ref: "#/components/responses/UnauthorizedError"

  /parser/history:
    get:
      security:
        - bearerAuth: [ ]
      summary: Return List of User Parser Queries History
      tags:
        - Parser
      operationId: getAllParserQueries
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParserResultOpenApi'
        '401':
          $ref: "#/components/responses/UnauthorizedError"


  /parser/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Return Parser Settings and It's Result File
      tags:
        - Parser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Parser Settings ID
      operationId: showParserSettings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParserResultOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
    post:
      security:
        - bearerAuth: [ ]
      summary: Start Parsing with Preset Settings with {id}
      tags:
        - Parser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Parser Settings ID
      operationId: runParser
      responses:
        '200':
          description: OK
        '400':
          description: This Parser Settings wasn't found
        '401':
          $ref: "#/components/responses/UnauthorizedError"

  /parser/{id}/download:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download Results File With the Specific ID
      tags:
        - Parser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: Parser Results File ID
      operationId: downloadFile
      responses:
        200:
          description: The file download was successful
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Internal Server Error

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Access token is missing or invalid

  schemas:

    UserOpenApi:
      type: object
      required:
        - username
        - email
        - userStatus
      properties:
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/RoleOpenApi"
        userStatus:
          type: object
          $ref: "#/components/schemas/UserStatusOpenApi"
        userParseSettings:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/UserParserSettingsOpenApi"
        parserResults:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/ParserResultOpenApi"
        telegramUserId:
          type: string

    RoleOpenApi:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: object
          $ref: "#/components/schemas/ERoleOpenApi"


    ERoleOpenApi:
      type: string
      enum:
        - ROLE_USER
        - ROLE_ADMIN

    UserStatusOpenApi:
      type: string
      enum:
        - CONFIRMED_ACCOUNT
        - WAIT_FOR_EMAIL_VERIFICATION

    LoginRequestOpenApi:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    SignupRequestOpenApi:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string

    JwtResponseOpenApi:
      type: object
      required:
        - accessToken
        - id
        - username
      properties:
        accessToken:
          type: string
          pattern: ^Bearer\s.+$
        id:
          type: integer
          format: int64
        username:
          type: string

    ParserResultOpenApi:
      type: object
      required:
        - id
        - userParserSettings
        - linkToDownloadResults
      properties:
        id:
          type: integer
          format: int64
        userParserSettings:
          type: object
          $ref: "#/components/schemas/UserParserSettingsOpenApi"
        linkToDownloadResults:
          type: string

    UserParserSettingsOpenApi:
      type: object
      required:
        - firstPageUrl
        - numOfPagesToParse
        - className
        - tagName
        - cssSelectorNextPage
        - header
        - parseSettingElements
        - outputFileType
      properties:
        firstPageUrl:
          type: string
        numOfPagesToParse:
          type: integer
          format: int32
        className:
          type: string
        tagName:
          type: string
        cssSelectorNextPage:
          type: string
        header:
          type: array
          items:
            type: string
        parseSettingElements:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/ElementLocatorOpenApi"
        outputFileType:
          type: object
          $ref: "#/components/schemas/OutputFileTypeOpenApi"


    ElementLocatorOpenApi:
      type: object
      required:
        - elementType
        - pathToLocator
      properties:
        id:
          type: integer
          format: int64
        elementType:
          type: object
          $ref: "#/components/schemas/ElementTypeOpenApi"
        pathToLocator:
          type: string
        extraPointer:
          type: string

    OutputFileTypeOpenApi:
      type: string
      enum:
        - XLSX
        - CSV

    ElementTypeOpenApi:
      type: string
      enum:
        - XPATH
        - TAG_ATTR
        - CSS